requirements.txt
numpy
scikit-image
h5py
Pillow
matplotlib
scipy
streamlit